+++ doc/whatsnew.rst
@@ -4,7 +4,9 @@ What's new
 Next release
 ============
 
-- Fix the nexus/cooling function and adding test for checking some input data (:pull:`236`).
+- Connect the water module to the cost module for cooling technologies (:pull:`245`).
+- Make setup of constraints for cooling technologies flexible and update solar csp tech. name (:pull:`242`). 
+- Fix the nexus/cooling function and add test for checking some input data (:pull:`236`).
 - Add :doc:`/project/circeular` project code and documentation (:pull:`232`).
 - Update water availability data and major code editing to allow a new test suite for the water module (:pull:`106`).
 - Expand :doc:`repro` with sections on :ref:`repro-doc` and :ref:`versioning`, including :ref:`a list of external model names and ‘versions’ <model-names>` like “MESSAGEix-GLOBIOM 2.0” (:issue:`224`, :pull:`226`).
@@ -22,6 +24,9 @@ Next release
 - Improve and extend :doc:`/material/index` (:pull:`218`).
 
   - Release of MESSAGEix-Materials 1.1.0 (:doc:`/material/v1.1.0`).
+Improve and extend :doc:`/material/index` (:pull:`218`).
+
+  - Release of MESSAGEix-Materials 1.1.0 (:doc:`/material/v1.1.0`)
 
 v2024.8.6
 =========